+++ src/main/java/com/arcbees/bourseje/client/admin/AdminModule.java
@@ -14,16 +14,16 @@
  * the License.
  */
 
-package com.arcbees.bourseje.client.realtime;
+package com.arcbees.bourseje.client.admin;
 
-import com.arcbees.bourseje.client.realtime.candidates.CandidatesModule;
-import com.arcbees.bourseje.client.realtime.endofvote.EndOfVoteModule;
-import com.arcbees.bourseje.client.realtime.howtovote.HowToVoteModule;
-import com.arcbees.bourseje.client.realtime.numberofvote.NumberOfVoteModule;
-import com.arcbees.bourseje.client.realtime.result.ResultModule;
+import com.arcbees.bourseje.client.admin.candidates.CandidatesModule;
+import com.arcbees.bourseje.client.admin.endofvote.EndOfVoteModule;
+import com.arcbees.bourseje.client.admin.howtovote.HowToVoteModule;
+import com.arcbees.bourseje.client.admin.numberofvote.NumberOfVoteModule;
+import com.arcbees.bourseje.client.admin.result.ResultModule;
 import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
 
-public class RealtimeModule extends AbstractPresenterModule {
+public class AdminModule extends AbstractPresenterModule {
     @Override
     protected void configure() {
         install(new HowToVoteModule());
@@ -32,7 +32,7 @@ public class RealtimeModule extends AbstractPresenterModule {
         install(new ResultModule());
         install(new EndOfVoteModule());
 
-        bindPresenter(RealtimePresenter.class, RealtimePresenter.MyView.class, RealtimeView.class,
-                RealtimePresenter.MyProxy.class);
+        bindPresenter(AdminPresenter.class, AdminPresenter.MyView.class, AdminView.class,
+                AdminPresenter.MyProxy.class);
     }
 }
