+++ src/main/java/com/arcbees/bourseje/client/admin/numberofvote/NumberOfVoteView.ui.xml
