+++ src/main/java/com/arcbees/bourseje/client/admin/result/ResultModule.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2014 ArcBees Inc.
+ * Copyright 2015 ArcBees Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,21 +14,14 @@
  * the License.
  */
 
-package com.arcbees.bourseje.client.realtime.result;
+package com.arcbees.bourseje.client.admin.result;
 
-import com.google.gwt.uibinder.client.UiBinder;
-import com.google.gwt.user.client.ui.Widget;
-import com.gwtplatform.mvp.client.ViewImpl;
+import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
 
-import javax.inject.Inject;
-
-public class ResultView extends ViewImpl implements ResultPresenter.MyView {
-    interface Binder extends UiBinder<Widget, ResultView> {
-    }
-
-    @Inject
-    ResultView(
-            Binder uiBinder) {
-        initWidget(uiBinder.createAndBindUi(this));
+public class ResultModule extends AbstractPresenterModule {
+    @Override
+    protected void configure() {
+        bindPresenter(ResultPresenter.class, ResultPresenter.MyView.class,
+                ResultView.class, ResultPresenter.MyProxy.class);
     }
 }
