+++ src/main/java/com/arcbees/bourseje/client/admin/result/ResultView.java
@@ -1,5 +1,5 @@
-/*
- * Copyright 2015 ArcBees Inc.
+/**
+ * Copyright 2014 ArcBees Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,12 +14,21 @@
  * the License.
  */
 
-package com.arcbees.bourseje.client.realtime.numberofvote;
+package com.arcbees.bourseje.client.admin.result;
 
-import com.gwtplatform.mvp.client.UiHandlers;
+import com.google.gwt.uibinder.client.UiBinder;
+import com.google.gwt.user.client.ui.Widget;
+import com.gwtplatform.mvp.client.ViewImpl;
 
-public interface NumberOfVoteUiHandlers extends UiHandlers {
-    void onLoginClicked();
+import javax.inject.Inject;
 
-    void onStartVoteClicked();
+public class ResultView extends ViewImpl implements ResultPresenter.MyView {
+    interface Binder extends UiBinder<Widget, ResultView> {
+    }
+
+    @Inject
+    ResultView(
+            Binder uiBinder) {
+        initWidget(uiBinder.createAndBindUi(this));
+    }
 }
