+++ src/main/java/com/arcbees/bourseje/client/admin/result/ResultView.ui.xml
