+++ src/main/java/com/arcbees/bourseje/client/admin/ui/Footer.ui.xml
